LPS-184134 Merge all these PortletDisplayTemplateUtil into one under the API module portlet-display-template-api and apply
