+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-taglib/src/main/resources/META-INF/resources/template_renderer/init-ext.jspf
@@ -15,4 +15,4 @@
 --%>
 
 <%@ page import="com.liferay.dynamic.data.mapping.model.DDMTemplate" %><%@
-page import="com.liferay.dynamic.data.mapping.taglib.internal.util.PortletDisplayTemplateUtil" %>
+page import="com.liferay.portlet.display.template.util.PortletDisplayTemplateUtil" %>
