+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-taglib/src/main/resources/META-INF/resources/template_selector/init-ext.jspf
@@ -18,7 +18,6 @@
 page import="com.liferay.dynamic.data.mapping.model.DDMTemplate" %><%@
 page import="com.liferay.dynamic.data.mapping.service.DDMTemplateLocalServiceUtil" %><%@
 page import="com.liferay.dynamic.data.mapping.taglib.internal.security.permission.resource.DDMTemplatePermission" %><%@
-page import="com.liferay.dynamic.data.mapping.taglib.internal.util.PortletDisplayTemplateUtil" %><%@
 page import="com.liferay.dynamic.data.mapping.util.DDMNavigationHelper" %><%@
 page import="com.liferay.portal.kernel.language.LanguageUtil" %><%@
 page import="com.liferay.portal.kernel.language.UnicodeLanguageUtil" %><%@
@@ -31,4 +30,5 @@ page import="com.liferay.portal.kernel.util.HtmlUtil" %><%@
 page import="com.liferay.portal.kernel.util.PortalUtil" %><%@
 page import="com.liferay.portal.kernel.util.PortletKeys" %><%@
 page import="com.liferay.portal.kernel.util.Validator" %><%@
-page import="com.liferay.portlet.display.template.PortletDisplayTemplate" %>
+page import="com.liferay.portlet.display.template.PortletDisplayTemplate" %><%@
+page import="com.liferay.portlet.display.template.util.PortletDisplayTemplateUtil" %>
