+++ modules/apps/portlet-display-template/portlet-display-template-api/bnd.bnd
@@ -4,4 +4,5 @@ Bundle-Version: 4.0.5
 Export-Package:\
 	com.liferay.portlet.display.template,\
 	com.liferay.portlet.display.template.constants,\
-	com.liferay.portlet.display.template.exportimport.portlet.preferences.processor
+	com.liferay.portlet.display.template.exportimport.portlet.preferences.processor,\
+	com.liferay.portlet.display.template.util
