+++ modules/apps/roles/roles-admin-web/src/main/java/com/liferay/roles/admin/web/internal/display/context/EditRolePermissionsFormDisplayContext.java
@@ -39,9 +39,9 @@ import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.portlet.display.template.util.PortletDisplayTemplateUtil;
 import com.liferay.portlet.rolesadmin.search.ResourceActionRowChecker;
 import com.liferay.roles.admin.web.internal.group.type.contributor.util.GroupTypeContributorUtil;
-import com.liferay.roles.admin.web.internal.util.PortletDisplayTemplateUtil;
 import com.liferay.taglib.search.ResultRow;
 
 import java.util.ArrayList;
